pyproject.toml
[tool.poetry]
name = "mobilizon-reshare"
version = "0.1.0"
description = ""
authors = ["Simone Robutti <simone.robutti@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dynaconf = "^3.1.4"
tortoise-orm = "^0.17"
aiosqlite = "^0.16"
Jinja2 = "^2.11.3"
requests = "^2.25.1"
arrow = "^1.1.0"
click = "^8.0.1"
beautifulsoup4 = "^4.9.3"
markdownify = "^0.9.2"

[tool.poetry.dev-dependencies]
responses = "^0.13.3"
pytest-asyncio = "^0.15.1"
asynctest = "^0.13.0"
pytest = "6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mobilizon-reshare="mobilizon_reshare.cli.cli:mobilizon_reshare"
